# src/main/java/com/probestack/forgestudio/design/api/ProjectMembersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.AddProjectMemberRequest;
import com.probestack.forgestudio.design.model.ProjectMember;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-06T23:23:39.454893700-07:00[America/Los_Angeles]")@Validated
@Tag(name = "Project Members", description = "the Project Members API")
public interface ProjectMembersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /projects/{projectId}/members : List project members
     *
     * @param projectId  (required)
     * @return Members list (status code 200)
     */
    @Operation(
        operationId = "projectsProjectIdMembersGet",
        summary = "List project members",
        tags = { "Project Members" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Members list", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ProjectMember.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/projects/{projectId}/members",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<ProjectMember>> projectsProjectIdMembersGet(
        @Parameter(name = "projectId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("projectId") UUID projectId
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "[ { \"role\" : \"role\", \"joinedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"userId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, { \"role\" : \"role\", \"joinedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"userId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" } ]";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * POST /projects/{projectId}/members : Add member to project
     *
     * @param projectId  (required)
     * @param addProjectMemberRequest  (required)
     * @return Member added (status code 201)
     */
    @Operation(
        operationId = "projectsProjectIdMembersPost",
        summary = "Add member to project",
        tags = { "Project Members" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Member added")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/projects/{projectId}/members",
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Void> projectsProjectIdMembersPost(
        @Parameter(name = "projectId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("projectId") UUID projectId,
        @Parameter(name = "AddProjectMemberRequest", description = "", required = true) @Valid @RequestBody AddProjectMemberRequest addProjectMemberRequest
    ) {
                        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                    }

}