# src/main/java/com/probestack/forgestudio/design/api/ProjectsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.CreateProjectRequest;
import com.probestack.forgestudio.design.model.Project;
import java.util.UUID;
import com.probestack.forgestudio.design.model.UpdateProjectRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-06T23:23:39.454893700-07:00[America/Los_Angeles]")@Validated
@Tag(name = "Projects", description = "the Projects API")
public interface ProjectsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /projects : Create project
     *
     * @param createProjectRequest  (required)
     * @return Project created (status code 201)
     */
    @Operation(
        operationId = "createProject",
        summary = "Create project",
        tags = { "Projects" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Project created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Project.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/projects",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Project> createProject(
        @Parameter(name = "CreateProjectRequest", description = "", required = true) @Valid @RequestBody CreateProjectRequest createProjectRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"specFormat\" : \"specFormat\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"baseUrl\" : \"baseUrl\", \"isArchived\" : true, \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"orgId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /projects/{projectId} : Get project
     *
     * @param projectId  (required)
     * @return Project details (status code 200)
     */
    @Operation(
        operationId = "getProject",
        summary = "Get project",
        tags = { "Projects" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Project details", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Project.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/projects/{projectId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Project> getProject(
        @Parameter(name = "projectId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("projectId") UUID projectId
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"specFormat\" : \"specFormat\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"baseUrl\" : \"baseUrl\", \"isArchived\" : true, \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"orgId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /projects : List projects
     *
     * @return List of projects (status code 200)
     */
    @Operation(
        operationId = "listProjects",
        summary = "List projects",
        tags = { "Projects" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of projects", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Project.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/projects",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Project>> listProjects(
        
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "[ { \"specFormat\" : \"specFormat\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"baseUrl\" : \"baseUrl\", \"isArchived\" : true, \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"orgId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }, { \"specFormat\" : \"specFormat\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"baseUrl\" : \"baseUrl\", \"isArchived\" : true, \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"orgId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" } ]";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * PATCH /projects/{projectId} : Update project
     *
     * @param projectId  (required)
     * @param updateProjectRequest  (optional)
     * @return Updated project (status code 200)
     */
    @Operation(
        operationId = "updateProject",
        summary = "Update project",
        tags = { "Projects" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Updated project", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Project.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/projects/{projectId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Project> updateProject(
        @Parameter(name = "projectId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("projectId") UUID projectId,
        @Parameter(name = "UpdateProjectRequest", description = "") @Valid @RequestBody(required = false) UpdateProjectRequest updateProjectRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"specFormat\" : \"specFormat\", \"createdAt\" : \"2000-01-23T04:56:07.000+00:00\", \"baseUrl\" : \"baseUrl\", \"isArchived\" : true, \"name\" : \"name\", \"description\" : \"description\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"orgId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}